begin reorganization
