+++ src/timing/frame_rate_limit.rs
@@ -3,7 +3,16 @@ use std::{
     time::{Duration, Instant},
 };
 
-use super::FrameRateLimit;
+/// A Vulkan application will generally run as fast as it possibly can to
+/// get images on screen. Often this is desirable, but when workloads are low
+/// it can cause unreasonably high frame-rates and therefore unexpectedly high
+/// CPU/GPU utilization. To prevent this, a frame rate limit can be imposed
+/// which just sleeps or yields for a bit of time each frame.
+pub struct FrameRateLimit {
+    frames_to_track: usize,
+    frame_starts: VecDeque<Instant>,
+    target_duration: Duration,
+}
 
 impl FrameRateLimit {
     /// Create a new frame rate limit for a given target fps
